requirements.txt
Django==6.0
djangorestframework==3.16.1
django-cors-headers==4.9.0
groq==0.11.0
python-dotenv==1.0.0

